Check a number can be expressed as a sum of primes
This program takes a positive integer as an input and checks if it can be expressed as a sum of prime numbers. If the integer can be expressed as a sum of primes then it outputs all the possibilities else it outputs a statement that the number cannot be expressed as a sum of primes.