Integer as a sum of primes
This program takes a positive integer as user input and checks if it can be expressed as a sum of prime numbers, if it can be expressed as a sum of primes it prints all different possibilities. If a number cannot be expressed as a sum of prime numbers, a statement is displayed to say the same.